Bug#32324234: MYSQL CRASH OPTIMIZE_DISTINCT_GROUP_ORDER
Summary:
While merging a derived table, all the expressions in the
outer query block that refer to the expressions from the
merged derived table, are replaced with view references
i.e a Item_view_ref object. However, the underlying
expression is shared with all the view references which
refer to the same expression in merged derived table.

While pushing a condition down to a materialized derived table,
expressions in the condition are replaced with derived
table expressions. However, if a condition has a view reference,
when the underlying expression is replaced with derived table
expressions, the other view references pointing to the same
expression in the outer query block will also reflect the
change. This is not desirable as the used tables information
for the outer query's view reference will not be correct leading
to problems later.

Solution:
Clone the underlying expression of a view reference before
it gets replaced with derived table expressions during
condition pushdown.

Porting notes: Can be dropped when porting forward to 8.0.26 or later

Reference patch: https://github.com/mysql/mysql-server/commit/215a4a730d2

Squash with null

Reviewed By: george-reynya

Differential Revision: D33970907

fbshipit-source-id: b16797de9cf