Fix before image when binlog entries are written during idempotent recovery
Summary:
When writes were failing during idempotent recovery (which is expected),
read set was not set to reflect the seconday's schema because
`mark_columns_per_binlog_row_image()` was not called in the force-write
to-binlog flow. Read set is used to generate before image in the binlog.

This is fine in most cases but can cause issues when the number of cols
in the secondary is greater than primary. In this case the BI contained
only the primary's cols. The issue doesn't affect MySQL since it can
deal with fewer cols in BI (that's what happens when row image is set to
MINIMAL) but mysql replicator lib expects the number of cols in BI to be
equal to number of cols in the table map event.

Reviewed By: yashtc

Differential Revision: D27928369

fbshipit-source-id: d33c8f21bd4