Upper HLC bound support
Summary: This diff adds support for "hlc_ts_upper_bound" query attribute. This attribute is expected to contain an HLC value which is used by mysqld to set an upper limit for HLC that will be associated with the given transaction. This implementation does not guarantee that the query will get an HLC below the set bound or fail. It'll fail the query in case if before starting the commit the DB HLC is already higher than the value specified by "hlc_ts_upper_bound" attribute. Though it's still possible that the query will succeed with HLC just slightly higher than "hlc_ts_upper_bound".

Reviewed By: abhinav04sharma

Differential Revision: D28003281

fbshipit-source-id: 25bc8f5619a