Backport fixes for Bug#32820437, Bug#32863713, Bug#32905044
Summary:
Bug#32820437: FULLTEXT MATCH CRASHED THE SERVER

When a condition is pushed down to a materialized derived
table, a clone of the derived table expression replaces
the field (from the outer query block) in the condition.
If this cloned item has a fulltext function, it needs to
be added to the ftfunc_list in the derived table query block.
However, while cloning, as the query block is not
set to the derived table query block, the fulltext function
gets added to the outer query block leading to problems
later.

Solution:
We now use derived query block to clone an item.

Bug#32863713: RECENT REGRESSION: CTE CRASH IN QUERY_BLOCK::MASTER_QUERY_EXPRESSION

To clone a derived table expression, we re-parse it. To re-parse the
expression, a new query block is created and used temporarily.
However, as part of an earlier fix, the query block to be used for parsing
was set to the query block provided in the context (derived table query
block or in case of a view ref, the merged derived table query block).
But, the query expression to use was still the query expression created
temporarily for parsing. Because of the mismatch between the query block
and the query expression that it belongs to, parser tries to access
an invalid query expression resulting in server exit.

We now set the query expression along with the query block to parse
the derived table expression.

Bug#32905044: CRASH AT CONDITION_PUSHDOWN::REPLACE_COLUMNS_IN_COND DURIN RQG CONCURRENCY RUNS

An earlier fix (Bug#32324234) made sure that a view reference needs to
be cloned before a condition is pushed down to the derived table.
To resolve the expression that was cloned, merged derived table's
context (query_block->first_context) was used. However, an Item_view_ref
object could also be created when a natural join column is created. For
this case, while resolving, we cannot be using the merged derived table's
context.
Also, query_block->first_context is updated when natural join is present
in the query. As a result, we can no more rely on first_context to give
the correct context for the merged derived table.

Solution:
While creating field translations for the table that is getting merged,
we capture the merged derived table's context. This context is then
saved to each of the Item_view_ref object created. We then use it
when view ref object is cloned.

Reference Patches:

https://github.com/mysql/mysql-server/commit/ca4d2ffa440ac
https://github.com/mysql/mysql-server/commit/28ec5fd1b
https://github.com/mysql/mysql-server/commit/18f05eff90

Porting notes: Can be dropped when rebasing to 8.0.26 or later

Squash with null

Reviewed By: lth

Differential Revision: D33983965

fbshipit-source-id: f290f8a8060