Add error message for purpose policy check fail
Summary:
This diff adds the error/warning message to be returned for purpose policy check fail. D28328304 is the diff that consumes this error message.

The diff also fixes an issue in sql_audit.cc which prevents the plugins to raise errors. This is based on the fix that went into 8.0: D14236916 (https://github.com/facebook/mysql-5.6/commit/be4696cc398313c3cabae53caaa5334501ab685a).

Reviewed By: lth

Differential Revision: D28288307

fbshipit-source-id: f45ded398f8