Fix buffer overrun when fetching response attributes
Summary:
The output of `mysql_resp_attr_find` is generally not nul terminated, as it refers directly to the buffer in the network packet. When `put_info` tries to print it out, it ends up printing past the length of the buffer. To fix, make the string nul terminated by duplicating it.

Also, fix a leak in the same function.

Squash with: D20186509 (https://github.com/facebook/mysql-5.6/commit/499a613f3610bcb9124ee5df74f5ad34e00ccd48)

Reviewed By: abal147

Differential Revision: D28423026

fbshipit-source-id: fa69c10da20