Move Plugin_connection_handler destructor to end_functions()
Summary:
Recent change D28608269 now causes ~Plugin_connection_handler() to be called after the plugin is unloaded so it crashes when tries to call `m_functions->end()`. The fix is to remove the destructor completely, and reintroduce `end()` call at the time of `my_connection_handler_reset()`.

Also add a trivial test that loads the plugin in the disabled state and never enables it to validate that the plugin can cleanly shutdown in this case.

Squash with D28608269 and D28250576.

Reviewed By: lth

Differential Revision: D28629099

fbshipit-source-id: 1c371e5544d