Fix heap overflow in group_relay_log_name handling
Summary:
We were accessing group_relay_log_name in
Query_log_event::do_apply_event_worker() but it's assigned only after
the coordinator thread encounters an end event (i.e. xid event or a
query event with "COMMIT" or "ROLLBACK" query). This was causing a race
between accessing group_relay_log_name in the worker thread and writing
it on the coordinator thread. We don't need to set transaction position
in events other than end event, so now we set transaction
position in query event only if it's an end event. The race is
eliminated because group_relay_log_name is set before enqueuing the
event to the worker thread (in both dep repl and vanilla mts).

Reviewed By: lth

Differential Revision: D28767273

fbshipit-source-id: 81fa0d6daac