MyRocks crashes on pure virtual function when querying rocksdb_trx list
Summary:
Rdb_transactions are added and removed from the global
list by the Rdb_transaction constructor/destructors.

However, information_schema.rocksdb_trx queries scan
this global list and calls the virtual function is_writebatch_trx()
to determine how to cast the object.

It is possible for the subclasses destructors to have completed, but
before Rdb_transaction's destructor can remove the object from the
global list that an information_schema.rocksdb_trx query is initiated.
The query walks the list of transactions and comes across a partially
deconstructed Rdb_transaction_impl object. When it calls
is_writebatch_trx() on this object, it triggers a pure function access
crash because the function has already been removed/cleared.

Fix this by adding the objects to the global list after
the object is fully constructed and remove it from the global list
when the object enters the destructor.

Reviewed By: yizhang82

Differential Revision: D29164519

fbshipit-source-id: aaf39f330c8