Bug #31685095 INNODB] ASSERTION FAILURE: FSP0FSP.CC:4648:NOT_FULL_N_USED_1 == NOT_FULL_N_USED_
Summary:
Each file segment inode contains a field FSEG_NOT_FULL_N_USED, which provides
information about the number of used pages in extents belonging to the
FSEG_NOT_FULL list of extents.  There is an assert which traverses the
FSEG_NOT_FULL list and calculates the number of used pages in this list
and then compares this value with the FSEG_NOT_FULL_N_USED.  This assert
failed sporadically.

Porting notes: Backported

https://github.com/mysql/mysql-server/commit/415c4aec0
https://github.com/mysql/mysql-server/commit/f18c71ff8
https://github.com/mysql/mysql-server/commit/742f14864

from upstream 8.0.27 to resolve lob::purge() issues. These can be
dropped when porting to 8.0.27 or later.

Reviewed By: george-reynya

Differential Revision: D34346144

fbshipit-source-id: 2ba7311b3c7