Generate code coverage using suite option
Summary:
This diff adds the ability to generate code coverage reports with "--suites" option to mysqltest.sh. The detailed changes are:
* Disable '--mem' option in coverage mode since this disables generating the code coverage profile files (*.profraw).
* When running a lot of tests, the code coverage profile files get purged when the server gets restarted in between the tests. Due to this, the coverage information of some of tests will be lost. This diff fixes it by moving the *.profraw files to a separate folder before the purge. A new directory is created in <mysql_build_dir>/coverage_files where these profile files are staged and purged at the end of the coverage run.
* Clean up of all the code coverage profile files at the end of the coverage run.

D29301764 contains the changes required in MySQL tools repos. D29301749 is the corresponding changes for MySQL 8.0.20. One limitation of the 5.6.35 implementation is: there is no support to run suites that belong to both innodb and rocksdb in the same mysqltest command. All the suites should either belong to innodb or all of them should belong to rocksdb.

Reviewed By: yizhang82

Differential Revision: D29333584

fbshipit-source-id: 476d75e17c9