Fix locking in partial indexes
Summary:
Fix a race condition during deletes in partial indexes. Currently we're not locking the prefix during deletes, which means that we could end up materializing deleted rows. For example, if we have:
1. T1 deletes a row A without committing
2. T2 reads row A during materialization, and generates SK for row A
3. T1 commits
4. T2 commits

Then the partial index will contain an extra row that shouldn't exist.

The solution is to also add locking there (similar to insert path).

Fix the stress test to catch this case, since previously, it was not surfacing errors correctly.

Also, create a new iterator object during materialization, since it is possible that the old iterator may be on an old superversion, and may not see all keys.

Reviewed By: hermanlee

Differential Revision: D29683427

fbshipit-source-id: ad8c1b7d4e9