Fix master_uuid in Raft follower (mysql)
Summary:
Slave used to connect to master to get that info in semi-sync mode. However, in raft mode, there is no direct communication between slave-master in mysql, and it is not possible (has host-port info but no user,password info).

Summary of changes: introduce master uuid attribute in RaftListenerCallbackArg (mysql and multitenancy header). Pass the master uuid via RaftListenerCallbackType::CHANGE_MASTER (in addition to host and port), and write the info in raft_change_master() after raft handler gets it.

Reviewed By: yashtc

Differential Revision: D29810744

fbshipit-source-id: 7f15c35cb29