Adding 2 log lines to better figure out errors in raft
Summary:
In some rare cases we are hitting an error where
the applier log is not hot and the next log is not found.
We want to print what the actual error code is there.

In second case we also want to know whether rli_init_info is
being called when rli is already inited.

Reviewed By: bhatvinay

Differential Revision: D29953871

fbshipit-source-id: 519799b3270