Fix rpl.rpl_io_thd_wait_for_disk_space
Summary:
If `THD::thd_enter_stage` is called from the signal handler thread, then current_thd would be nullptr, resulting in a crash. This is similar to the fix in D29768712, and I missed this other callsite.

That said, this code path is only entered when the test timeouts out (and hence the server gets a kill signal). Also increase the valgrind timeout to help the test pass.

Squash with: D29768712

Reviewed By: hermanlee

Differential Revision: D30173407

fbshipit-source-id: db1651a54f7