work around raft not having clean shutdown
Summary:
Even on clean shutdown, raft can make it so that some trxs get aborted
in the engine (but remain in the binlog file). During clean shutdown, the
binlog file is marked as closed and recovery used to depend on this
'feature' to assume that the closed binlog file will not contain any uncommitted
trxs. This diff makes it so that this dependency is removed and we
always start mysqld as if we are recovering the last binlog file.
This will allow raft plugin write a rotate event into this binlog file when it
rejoins the ring.

Reviewed By: abhinav04sharma

Differential Revision: D30227164

fbshipit-source-id: f39e7f124b5