Add TABLE_INSTANCE to COLUMN_STATISTICS
Summary:
When there are multiple instances of the same table in a single query, we did not distinguish between different instances of the same table. This diff adds a column called TABLE_INSTANCE which is going to be unique for every instance of a table
```
(SQL_ID, TABLE_SCHEMA, TABLE_NAME, TABLE_INSTANCE)
```

This would be used for creation of better suggestions in case of self joins. `TABLE_INSTANCE` is supposed to be an identifier for a specific instance within a query. There is no need for it to be ordered. The only property we are going to use is distinctness.

Reviewed By: satya-valluri, mzait

Differential Revision: D30417495

fbshipit-source-id: 6723326b009