Use sentinel value to increase write performance
Summary:
Currently, the check for materialized rows for partial indexes is done by seeking to the prefix of the group and checking to see if any rows are present. This seek can be expensive, and can lead to replication lag on the write path, and extra cpu cost on the read path.

To avoid this, we can write a sentinel value with the key as the prefix, and an empty value. Instead of seeking to check for materialized rows, the existence of this key will imply that the group is materialized.

There are a few side effects from this change:
- Previously, if all rows in a group are deleted, then the group can become unmaterialized. After this change though, once a group is materialized, it will always remain materialized even after deletes, because nothing ever deletes the sentinel value.
- Although read paths that use the iterator interface are instrumented to ignore sentinel values, we still have reads coming from compaction that do not use our iterators. There's no easy way for compaction to ignore these values (since compaction may run before the `Rdb_key_def` struct is fully initialized), so it's possible that stats are somewhat inaccurate (although they're already inaccurate with partial indexes). Related to this is that TTL and partial indexes cannot be used together.

Reviewed By: hermanlee

Differential Revision: D30619361