SHOW INNODB ENGINE STATUS show query digest for deadlock queries
Summary:
We had show_query_digest that shows query digest and skips physical records in most cases for SHOW ENGINE STATUS already for sensitive productions, but it doesn't work for deadlock detection part of the output.

The reason is deadlock information are collected in the background in a file whenever deadlock is happening, so when you set show_query_digest=1 and run `SHOW ENGINE STATUS` it is already too late. Besides creating data structures that gives proper abstraction, the easiest way to fix this is to keep two separate file - one for the regular case and the other for query digest for sensitive queries, and we hand out whichever file as needed.

Also back ported lock_wait_will_wait from 8.0 to 5.6 to make the test case stable for innodb.

Reviewed By: luqun

Differential Revision: D30845463

fbshipit-source-id: ce5fba58de8