Fix BKA when using RocksDB MultiGet
Summary:
Currently, when we call into MultiGet during MRR, the sorted parameter is determined by seeing if the key is a primary key or not.

While this works for usages of MRR via range plans, it does not work when MRR is used via BKA. This is because although we are doing PK lookups, the driving table in the join may not have their rows sorted by the PK.

The fix is to check the `HA_MRR_NO_ASSOCIATION` flag which is only set when MRR is used via the range plan codepath, and not BKA. The only time when the keys are sorted if we are doing PK lookups via range plans.

There is one additional bug here, which is the fact that we would still use sorted MultiGet if we had a PK in a reverse cf. This would also produce undefined behaviour as we would be passing the keys in the wrong order.

Backported from 8.0

Reviewed By: yoshinorim

Differential Revision: D30884227

fbshipit-source-id: 1fabdb301f3