Include async_id as a valid attribute to parse
Summary:
This makes it possible to pass async_id as a query attribute, and for it to show up as client attributes.

This meant to replace the current mechanism of passing this via query comments.

Porting Notes:

This is only needed until migration from query comments to query attributes are complete. Before that we need to parse both query comments and query / connection attributes.

The only change I'd need to tweak is to serialize_client_attrs happens before alloc_query, so this changes serialize_client_attr and explicitly pass the query + length.

Reviewed By: lth

Differential Revision: D31329760

fbshipit-source-id: 4210ddd513b