avoid crashing the server when purging invalid binlog files
Summary:
Trying to purge a binlog file which is invalid (either invalid name or
is already purged before) can end up crashing the server when
binlog_error_action is set to ABORT_SERVER. Crashing only make sense
when the error is because of some file system operation (disk full,
corruption etc). This diff fixes the problem by avoiding the crash when
the binlog filename is not found in the index file.

Reviewed By: abhinav04sharma

Differential Revision: D31154700

fbshipit-source-id: f82dfb59154