Take SQL statistics snapshot at periodic intervals
Summary:
The SQL statistics are captured today at the end of the execution of a SQL statement. This has two limitations:
a) we do not have details of resource usage of long running queries until they are done
b) the capture into hive is misleading: a long running SQL statement shows in the top of a capture period at the end even though it did not contribute during that time window

With this change, allowing for more frequent stat collection during query execution. The server variable: min_examined_row_count_sql_stats will be used to determine the sampling intervals.

The entire tracking is stored in the thread structure.

The collection of query_sample is still in conjunction with the variable: max_digest_sample_age

Reviewed By: george-reynya

Differential Revision: D31601612

fbshipit-source-id: f90d9329c7b