Disable the clock_gettime call by default for optimizer cputime pfs metrics
Summary:
Every call to clock_gettime(CLOCK_THREAD_CPUTIME_ID) makes a system call which consumes lot of cpu cycles and this is not optimized for vDSO. Disabling the SUM_COMPILATION_CPU and SUM_INDEX_DIVE_CPU by default. If we want to gather those stats, then we can selectively enable that.

%CPU Savings: ~0.5%

Reviewed By: yizhang82

Differential Revision: D31676544