Make the type of argument to the query attribute 'response_attrs_contain_server_cpu' an integer
Summary:
Squash with D31626851".

Change the type of argument to the query attribute 'response_attrs_contain_server_cpu' to an integer. This avoids type conversions while checking the query attribute.

Note that the response attribute will be considered enabled only the value is set to '1'. It will be considered as disabled otherwise.

Reviewed By: yoshinorim

Differential Revision: D31734051

fbshipit-source-id: 3f8e620edd6